docs: ✏️ custom triggers & provide hooks to do that
✅ Closes: #705
