+++ packages/mui-layout/src/components/SidebarTrigger/SidebarTrigger.tsx
@@ -4,9 +4,7 @@ import IconButton, { IconButtonProps } from '@material-ui/core/IconButton';
 import ArrowLeft from '@material-ui/icons/KeyboardArrowLeftRounded';
 import ArrowRight from '@material-ui/icons/KeyboardArrowRightRounded';
 import MenuRounded from '@material-ui/icons/MenuRounded';
-import EdgeTriggerCompiler from '../../compilers/EdgeTriggerCompiler';
-import { useSidebarCta } from '../../hooks';
-import { createDisplayNone } from '../../utils';
+import { useSidebarTrigger } from '../../hooks';
 import { CtaProps } from '../../types';
 import { generateStyledProxyCreator } from '../Shared/StyledProxy';
 
@@ -41,11 +39,11 @@ export default (styled: any) => {
     const {
       id,
       anchor,
-      breakpoints,
-      edgeSidebar,
+      hiddenStyles,
       state,
       setOpen,
-    } = useSidebarCta(sidebarId, 'SidebarTrigger');
+    } = useSidebarTrigger(sidebarId, 'SidebarTrigger');
+    console.log('hiddenStyles', hiddenStyles);
     const classes = useStyles({ ...props, anchor });
     const getArrow = () => {
       if (!state) return null;
@@ -54,14 +52,10 @@ export default (styled: any) => {
       if (anchor === 'right') return <ArrowRight {...SvgIconProps} />;
       return null;
     };
-    const styles = createDisplayNone(
-      EdgeTriggerCompiler(edgeSidebar).getHiddenBreakpoints(id),
-      breakpoints
-    );
     return (
       <StyledIconBtn
         {...props}
-        styles={styles}
+        styles={hiddenStyles}
         classes={classes}
         onClick={(e: React.MouseEvent<HTMLButtonElement>) => {
           if (typeof onClick === 'function') onClick(e);
